Rhubarb for BitBucket
I came across tpope/vim-rhubarb#61 and thought "There's no way the lift
for this would be heavy". So I thought to give it a stab.

This is the introductory commit for supporting Bitbucket with Rhubarb.
One does not need to do any extranneous configuration changes to use
this. One just needs to have the github_enterprise_urls set to BitBucket

    g:github_enterprise_urls = ['https://bitbucket.org']

There are s few assumptions here.

- Bitbucket uses /src/ in place og GitHub's /blob/
- Bitbucket uses #lines- in place of Github's #L